Move agents based on distance to cell centroid. Create live plot
New parameters have been added as well: An ID to
Agent class, a time step for the movement and simulation, and a Kp
for getting the velocity.

It was fixed a bug causing an intersection to dissapear as it was
going beyond the borders by an infinitesimal value. A tolerance
has been added to the comparison of whether the points are inside
the map.