Fix integral calculus. Coverage control works in simulation (#6)
The coverage control integrates a Gaussian function representing
the density of information:

1. scipy.spatial.Delaunay is used for triangulation of the cells.
The result is printed during the simulation.

2. For each obtained triangle, it is calculated the Jacobian for
tranforming it into a right angle between [0,0], [0,1], and [1,0]

3. The mass and centre of mass of each triangle is obtained by
integrating within the transformed triangle, using the pullback of
the density function in the transformed space.

4. The total mass and centroid of the cell is obtained by
superposition of the centroids of the obtained triangles.